fix: moving dynamodb-tables back, deleting other files
